Create actions to publish API OAS beta docs
Re-reverting this with single quotes for cron tabs. I doubt it fixes it but not sure what else to try at this point as the scheduled actions are running at the wrong times.

This PR creates three actions:

One that runs weekly (except for first week of month) to publish a new pre-release release (with pre-release version bumped) with updated spec2.json and spec3.json
One that runs monthly (first Wednesday of the month) to publish a new pre-release release (with minor version bumped) with updated spec2.json and spec3.json
One that runs weekly (on every Wednesday) to publish a new GA release (with minor version bumped) with updated spec2.json and spec3.json
Both of these jobs can also be triggered manually.

Both of these are pulling from our AWS s3 bucket that house the latest version of our OAS docs